requirements.txt
Appium-Python-Client==1.1.0
behave==1.2.6
parse==1.19.0
parse-type==0.5.2
selenium==3.141.0
six==1.15.0
urllib3==1.26.4
